rset: Introduce `checksumFrom` for triggering rollout restarts
