+++ docs/api/v1/resourceset.md
@@ -561,6 +561,78 @@ spec:
         cluster: << inputs.cluster >>
 ```
 
+#### Computing checksums from ConfigMaps and Secrets
+
+To trigger a rolling update of a workload when the data of a referenced
+ConfigMap or Secret changes, the `fluxcd.controlplane.io/checksumFrom` annotation
+can be placed on any `metadata.annotations` map in the ResourceSet template.
+
+The annotation value must be a comma-separated list of references in the format
+`Kind/namespace/name`, where `Kind` is either `ConfigMap` or `Secret`. 
+
+During reconciliation, the operator resolves each reference, computes a SHA256
+digest over the combined data, and writes the result into a sibling
+`fluxcd.controlplane.io/checksum` annotation as `sha256:<hex>`. The original
+`checksumFrom` annotation is preserved.
+
+Example of using `checksumFrom` to trigger a Pod rollout on a Deployment
+whenever the referenced ConfigMap or Secret changes:
+
+```yaml
+spec:
+  inputs:
+    - name: podinfo
+  resources:
+    - apiVersion: apps/v1
+      kind: Deployment
+      metadata:
+        name: << inputs.name >>
+        namespace: << inputs.provider.namespace >>
+      spec:
+        selector:
+          matchLabels:
+            app.kubernetes.io/name: << inputs.name >>
+        template:
+          metadata:
+            annotations:
+              fluxcd.controlplane.io/checksumFrom: |
+                ConfigMap/<< inputs.provider.namespace >>/<< inputs.name >>-config,
+                Secret/<< inputs.provider.namespace >>/<< inputs.name >>-secret
+            labels:
+              app.kubernetes.io/name: << inputs.name >>
+          spec:
+            containers:
+              - name: podinfo
+                image: ghcr.io/stefanprodan/podinfo:6.0.0
+                envFrom:
+                  - configMapRef:
+                      name: << inputs.name >>-config
+                  - secretRef:
+                      name: << inputs.name >>-secret
+```
+
+After reconciliation, the Deployment's pod template will carry both annotations:
+
+```yaml
+template:
+  metadata:
+    annotations:
+      fluxcd.controlplane.io/checksumFrom: "ConfigMap/apps/podinfo-config,Secret/apps/podinfo-secret"
+      fluxcd.controlplane.io/checksum: "sha256:abc123def456..."
+```
+
+When the data of any referenced ConfigMap or Secret changes, the computed
+checksum changes, the pod template hash changes, and Kubernetes performs a
+rolling update of the Deployment.
+
+References to ConfigMaps or Secrets generated by the same ResourceSet are
+resolved from the pending apply. For references to resources outside the ResourceSet,
+the checksum refreshes on the next reconciliation of the ResourceSet.
+
+Note that on [multi-tenant clusters](#role-based-access-control), the service account
+used by the ResourceSet must have the necessary permissions to read the ConfigMaps
+and Secrets referenced by `checksumFrom`, if those are outside the ResourceSet itself.
+
 #### Conditional resource exclusion
 
 To exclude a resource based on input values, the `fluxcd.controlplane.io/reconcile` annotation can be set
