+++ internal/entitlement/preflight_test.go
@@ -1,7 +1,7 @@
 // Copyright 2025 Stefan Prodan.
 // SPDX-License-Identifier: AGPL-3.0
 
-package builder
+package entitlement
 
 import (
 	"testing"
